--- game/src/main/react-map/src/MapGrid/MapGridRECOVER.tsx
@@ -1,105 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import style from './MapGridStyle.module.css';
-import { Player } from './Player';
-import redImage from '../images/Character_Movement/Red_Standing_South.png';
-import TestImage from "./TestImage";
-import purple from "../images/Character_Movement/Purple.png";
-
-interface MapGridProps {
-    currentPlayer: Player;
-    otherPlayers: Player[];
-}
-
-const MapGrid: React.FC<MapGridProps> = ({ currentPlayer, otherPlayers }) => {
-    const [gridKey, setGridKey] = useState(0);
-
-    useEffect(() => {
-        // Update the key whenever currentPlayer or otherPlayers change
-        setGridKey(prevKey => prevKey + 1);
-    }, [currentPlayer, otherPlayers]);
-
-    const grid: any[][] = [
-        ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
-        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
-        ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
-    ];
-
-    const radius = 5;
-
-    // Calculate sub-grid boundaries based on player position
-    const startRow = Math.max(0, currentPlayer.getY() - radius);
-    const endRow = Math.min(grid.length - 1, currentPlayer.getY() + radius);
-    const startCol = Math.max(0, currentPlayer.getX() - radius);
-    const endCol = Math.min(grid[0].length - 1, currentPlayer.getX() + radius);
-
-    // Center View
-    const paddingTop = Math.max(0, radius - (currentPlayer.getY() - startRow)) * 30;
-    const paddingLeft = Math.max(0, radius - (currentPlayer.getX() - startCol)) * 30;
-
-    // const TestImage = () => (
-    //     <img src={redImage} alt="Red Standing South" />
-    // );
-
-    return (
-        <div key={gridKey} className={style.root} style={{ paddingTop, paddingLeft }}>
-            {grid.slice(startRow, endRow + 1).map((row, rowIndex) => (
-                <div key={rowIndex} className={style.row} >
-                    {row.slice(startCol, endCol + 1).map((cell, colIndex) => {
-                        // Check if any other player is at this position
-                        const otherPlayer = otherPlayers.find(player => player.getX() === startCol + colIndex && player.getY() === startRow + rowIndex);
-                        let cellContent = cell;
-                        if (otherPlayer) {
-                            // If there's another player, use "@" symbol with player's color
-                            const playerColor = otherPlayer.getColor();
-                            cellContent = (
-                                <span style={{ color: playerColor }}>{'@'}</span>
-                            );
-                        } else if (startRow + rowIndex === currentPlayer.getY() && startCol + colIndex === currentPlayer.getX()) {
-                            // If it's the current player, use "@" symbol with current player's color
-                            const playerColor = currentPlayer.getColor(); // Spielerfarbe
-                            const playerImage = '../images/Character_Movement/Purple.png';
-
-                            cellContent = (
-                                <span style={{
-                                    backgroundImage: `url(${purple})`,
-                                    backgroundSize: 'cover',
-                                    width: '25px',
-                                    height: '25px'
-                                }}>
-                        </span>
-                            );
-                        }
-                        return (
-                            <span key={colIndex} className={style.cell}>
-                                {cellContent}
-                            </span>
-                        );
-                    })}
-                </div>
-            ))}
-        </div>
-    );
-};
-
-export default MapGrid;
