+++ game/src/main/react-map/src/MapGrid/MapGridRECOVERY2.tsx
@@ -1,0 +1,411 @@
+import {Player} from "./Player";
+import test from './Lobby.png';
+// import test3 from '../images/Maps/Lobby.tmx';
+import React, {useEffect, useRef, useState, useCallback, useMemo} from 'react';
+import {TiledResource} from "@excaliburjs/plugin-tiled";
+import lobbyImage from '../images/Maps/Lobby.png';
+import style from "./MapGridStyle.module.css";
+import purple from "../images/Character_Movement/Purple.png";
+import red from "../images/Character_Movement/red.jpg";
+import blue from "../images/Character_Movement/Blue.jpg";
+import green from "../images/Character_Movement/Green.jpg";
+import orange from "../images/Character_Movement/Orange.jpg";
+import yellow from "../images/Character_Movement/Yellow.png";
+import black from "../images/Character_Movement/Black.jpg";
+import white from "../images/Character_Movement/White.jpg";
+import brown from "../images/Character_Movement/Brown.jpg";
+import cyan from "../images/Character_Movement/Cyan.jpg";
+import lime from "../images/Character_Movement/Lime.jpg";
+import pink from "../images/Character_Movement/Pink.jpg";
+import dead from "../images/Character_Movement/dead.png";
+
+import skeldImage from '../images/Maps/Skeld.png';
+import '../CSS/MapGrid.css';
+
+interface MapGridProps {
+    currentPlayer: any;
+    otherPlayers: any[];
+}
+
+const colorToImageUrl = {
+    purple: purple,
+    red: red,
+    blue: blue,
+    green: green,
+    orange: orange,
+    yellow: yellow,
+    black: black,
+    white: white,
+    brown: brown,
+    cyan: cyan,
+    lime: lime,
+    pink: pink,
+    dead: dead,
+};
+
+const MapGridRECOVERY2: React.FC<MapGridProps> = ({currentPlayer, otherPlayers}) => {
+
+    const [otherPlayer, setOtherPlayer] = useState(otherPlayers);
+    const [playerImage, setPlayerImage] = useState(colorToImageUrl[currentPlayer.getColor()]);
+    const [otherPlayerImages, setOtherPlayerImages] = useState({});
+    const canvasRef = useRef<HTMLCanvasElement>(null);
+    const [scrollPosition, setScrollPosition] = useState({x: 0, y: 0});
+    const [gridKey, setGridKey] = useState(0);
+    const containerRef = useRef(null);
+
+
+    const [loadedImage, setLoadedImage] = useState(null);
+
+    useEffect(() => {
+        const image = new Image();
+        image.src = skeldImage;
+        image.onload = () => {
+            setLoadedImage(image);
+        };
+    }, [skeldImage]);
+
+
+    const playerImages = {
+        red: "../images/Character_Movement/red.jpg",
+        cyan: "../images/Character_Movement/Cyan.jpg",
+        lime: "../images/Character_Movement/Lime.jpg",
+        pink: "../images/Character_Movement/Pink.jpg",
+        gray: "../images/Character_Movement/Gray.jpg",
+        blue: "../images/Character_Movement/Blue.jpg",
+        green: "../images/Character_Movement/Green.jpg",
+        black: "../images/Character_Movement/Black.jpg",
+        white: "../images/Character_Movement/White.jpg",
+        brown: "../images/Character_Movement/Brown.jpg",
+        orange: "../images/Character_Movement/Orange.jpg",
+        yellow: "../images/Character_Movement/Yellow.jpg",
+        purple: "../images/Character_Movement/Purple.png",
+        dead: "../images/Character_Movement/dead.png",
+    };
+
+    useEffect(() => {
+        setPlayerImage(colorToImageUrl[currentPlayer.getColor()]);
+    }, [currentPlayer.getColor()]);
+
+    // const [currentPlayerImage, setCurrentPlayerImage] = useState(playerImages[currentPlayer.getColor()]);
+
+    // useEffect(() => {
+    //     setCurrentPlayerImage(playerImages[currentPlayer.getColor()]);
+    // }, [currentPlayer.getColor()]);
+    const currentPlayerImage = useMemo(() => {
+        return playerImages[currentPlayer.getColor()];
+    }, [currentPlayer.getColor()]);
+
+
+    useEffect(() => {
+        // Update the key whenever currentPlayer or otherPlayers change
+        setGridKey(prevKey => prevKey + 1);
+    }, [currentPlayer, otherPlayers]);
+
+
+    useEffect(() => {
+        otherPlayers.forEach(player => {
+            updatePlayerImage(player);
+        });
+    }, [otherPlayers]);
+
+    const updatePlayerImage = (player) => {
+        let newImage;
+        if (!player.getMovable()) {
+            newImage = colorToImageUrl['dead'];
+            newImage = lime;
+        } else if (player.getColor() === 'dead') {
+            newImage = colorToImageUrl['dead'];
+        } else {
+            // newImage = lime;
+            newImage = colorToImageUrl[player.getColor()];
+        }
+        setOtherPlayerImages(prevImages => ({...prevImages, [player.getSessionId()]: newImage}));
+    };
+
+
+    // const grid: any[][] = [
+    //     ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
+    //     ['#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#'],
+    //     ['#', '.', '.', '.', '.', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+    //     ['#', '.', '.', '.', '.', '#', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '#'],
+    //     ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+    //     ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '#'],
+    //     ['#', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '#'],
+    //     ['#', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '#', '.', '.', '#'],
+    //     ['#', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+    //     ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+    //     ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+    //     ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
+    // ];
+
+    // const grid: any[][] = [
+    //     [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 2, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 2, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 2, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+    //     [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+    // ];
+
+        const grid: any[][] = [
+        ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '?', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '#', '#', '#', '#', '#', '#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
+        ['#', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '.', '#', '#', '.', '.', '.', '.', '.', '#'],
+        ['#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#', '#'],
+    ];
+
+    const radius = 5;
+    const scaleFactor = 2;
+
+
+    // Calculate sub-grid boundaries based on player position
+    // const startRow = Math.max(0, currentPlayer.getY() - radius);
+    // const endRow = Math.min(grid.length - 1, currentPlayer.getY() + radius);
+    // const startCol = Math.max(0, currentPlayer.getX() - radius);
+    // const endCol = Math.min(grid[0].length - 1, currentPlayer.getX() + radius);
+    //
+    // // Center View
+    // const paddingTop = Math.max(0, radius - (currentPlayer.getY() - startRow)) * 30;
+    // const paddingLeft = Math.max(0, radius - (currentPlayer.getX() - startCol)) * 30;
+
+
+    // useEffect(() => {
+    //     const context = canvasRef.current?.getContext('2d');
+    //     if (context) {
+    //         const image = new Image();
+    //         image.src = skeldImage;
+    //
+    //         image.onload = () => {
+    //             // Transformation zurücksetzen, um sicherzustellen, dass keine vorherigen Transformationen angewendet werden
+    //             context.setTransform(1, 0, 0, 1, 0, 0);  // Setzt die Transformation auf die Standardeinstellung zurück
+    //             context.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height); // Löscht den Canvas
+    //
+    //             const scaleFactor = 2;
+    //             context.scale(scaleFactor, scaleFactor);  // Skalierung anwenden
+    //
+    //             // Position berechnen, um das Bild zentral zu positionieren, basierend auf der Spielerposition
+    //             // const imageXPosition = (canvasRef.current.width / 2) / scaleFactor - image.width / 2;
+    //             // const imageYPosition = (canvasRef.current.height / 2) / scaleFactor - image.height / 2;
+    //
+    //             const imageXPosition = (canvasRef.current.width / 2) / scaleFactor - currentPlayer.getX() * scaleFactor;
+    //             const imageYPosition = (canvasRef.current.height / 2) / scaleFactor - currentPlayer.getY() * scaleFactor;
+    //
+    //             // context.translate(imageXPosition, imageYPosition);
+    //             // context.drawImage(image, imageXPosition, imageYPosition);
+    //             context.drawImage(image, 0, 0, image.width, image.height, 0, 0, image.width, image.height);
+    //         };
+    //     }
+    // }, [currentPlayer, skeldImage]);
+
+    const calculateDistance = (x1: number, y1: number, x2: number, y2: number) => {
+        return Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2));
+    };
+
+    // return (
+    //     <div style={{position: 'absolute', overflow: 'scroll', width: '100%', height: '100%', left: 20, top: 200}}>
+    //         <canvas ref={canvasRef} id="game" width="2500" height="1800" style={{border: '1px solid black'}}></canvas>
+    //         <div key={gridKey} className={style.root} style={{position: 'absolute', top: 20, left: 180}}>
+    //             {grid.map((row, rowIndex) => (
+    //                 <div key={rowIndex} className={style.row}>
+    //                     {row.map((cell, colIndex) => {
+    //                         // Check if any other player is at this position
+    //                         const otherPlayer = otherPlayers.find(player => player.getX() === colIndex && player.getY() === rowIndex);
+    //                         let cellContent = cell;
+    //                         if (otherPlayer) {
+    //                             const playerImage = otherPlayerImages[otherPlayer.getSessionId()];
+    //                             return (
+    //                                 <span style={{
+    //                                     backgroundImage: `url(${playerImage})`,
+    //                                     backgroundSize: 'cover',
+    //                                     width: '25px',
+    //                                     height: '25px'
+    //                                 }}>
+    //                                 </span>
+    //                             );
+    //                         }
+    //                         else if (rowIndex === currentPlayer.getY() && colIndex === currentPlayer.getX()) {
+    //                             const playerColor = currentPlayer.getColor();
+    //                             console.log('playerColor ' + currentPlayer.getColor());
+    //
+    //                             cellContent = (
+    //                                 <span style={{
+    //                                     backgroundImage: `url(${playerImage})`,
+    //                                     backgroundSize: 'cover',
+    //                                     width: '25px',
+    //                                     height: '25px'
+    //                                 }}>
+    //                                 </span>
+    //                             );
+    //                         }
+    //                         return (
+    //                             <span key={colIndex} className={style.cell}>
+    //                             {cellContent}
+    //                         </span>
+    //                         );
+    //                     })}
+    //                 </div>
+    //             ))}
+    //         </div>
+    //     </div>
+    // );
+
+
+
+
+    useEffect(() => {
+        const context = canvasRef.current?.getContext('2d');
+        if (context) {
+            context.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height);
+            const image = new Image();
+            image.src = skeldImage;
+
+            image.onload = () => {
+                context.setTransform(1, 0, 0, 1, 0, 0);
+                context.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height);
+
+                const scaleFactor = 2;
+                context.scale(scaleFactor, scaleFactor);
+
+                // const imageXPosition = (canvasRef.current.width / 2) / scaleFactor - image.width / 2;
+                // const imageYPosition = (canvasRef.current.height / 2) / scaleFactor - image.height / 2;
+                const imageXPosition = (canvasRef.current.width / 2) / scaleFactor - currentPlayer.getX() * scaleFactor;
+                const imageYPosition = (canvasRef.current.height / 2) / scaleFactor - currentPlayer.getY() * scaleFactor;
+
+                // context.translate(imageXPosition, imageYPosition);
+                // context.drawImage(image, imageXPosition, imageYPosition);
+                context.drawImage(image, 0, 0, image.width, image.height, -0, 0, image.width, image.height);
+            };
+
+            const drawPlayer = (player, imageSrc) => {
+                const img = new Image();
+                img.src = imageSrc;
+                img.onload = () => {
+                    const targetWidth = 15;
+                    const targetHeight = 15;
+
+                    const xPosition = player.getX() * targetWidth;
+                    const yPosition = player.getY() * targetHeight;
+
+
+                    context.drawImage(img, xPosition + 80, yPosition + 7.5, targetWidth, targetHeight);
+                    // context.drawImage(img, player.getX() * 15, player.getY() * 15, 15, 15);
+                };
+            };
+
+            drawPlayer(currentPlayer, colorToImageUrl[currentPlayer.getColor()]);
+
+            otherPlayers.forEach(player => {
+                drawPlayer(player, colorToImageUrl[player.getColor()]);
+            });
+        }
+    }, [currentPlayer, skeldImage, otherPlayers]);
+
+    return (
+        <div className={style.root} style={{position: 'absolute', overflow: 'auto', overflowX: 'auto', overflowY: 'auto', width: '150%', height: '150%', top: 180, left: 20}}>
+            {/*<canvas ref={canvasRef} width="2500" height="1800" style={{border: '1px solid black'}}></canvas>*/}
+
+            <div key={gridKey} className={style.root} style={{position: 'absolute', top: 20, left: 180}}>
+                {grid.map((row, rowIndex) => (
+                    <div key={rowIndex} className={style.row}>
+                        {row.map((cell, colIndex) => {
+                            const otherPlayer = otherPlayers.find(player => player.getX() === colIndex && player.getY() === rowIndex);
+                            let cellContent = cell;
+                            if (otherPlayer) {
+                                const playerImage = otherPlayerImages[otherPlayer.getSessionId()];
+                                return (
+                                    <span style={{
+                                        backgroundImage: `url(${playerImage})`,
+                                        backgroundSize: 'cover',
+                                        width: '30px',
+                                        height: '30px'
+                                    }}>
+                                                </span>
+                                );
+                            } else if (rowIndex === currentPlayer.getY() && colIndex === currentPlayer.getX()) {
+                                const playerColor = currentPlayer.getColor();
+                                cellContent = (
+                                    <span style={{
+                                        backgroundImage: `url(${playerImage})`,
+                                        backgroundSize: 'cover',
+                                        width: '30px',
+                                        height: '30px'
+                                    }}>
+                                                 </span>
+                                );
+                            }
+                            return (
+                                <span key={colIndex} className={style.cell}>
+                                             {cellContent}
+                                          </span>
+                            );
+                        })}
+                    </div>
+                ))}
+            </div>
+
+
+
+        </div>
+        // </div>
+    );
+}
+
+
+// export default MapGridRECOVERY2;
+export default React.memo(MapGridRECOVERY2);
