+++ game/src/main/react-map/src/index.html
@@ -10,6 +10,6 @@
 <!--    <canvas id="canvasId"></canvas>-->
     <div id="root"></div>
     <script type="module" src="index.tsx"></script>
-<!--    <script src="./MapGrid/MapGrid.tsx"></script>-->
+<!--    <script src="./MapGrid/MapGridRECOVERY2.tsx"></script>-->
   </body>
 </html>
