+++ game/src/main/react-map/src/index.tsx
@@ -2,13 +2,13 @@ import "../init.js";
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import MapGrid from "./MapGrid/MapGrid";
+import MapGrid from "./MapGrid/MapGridRECOVERY2";
 import React from "react";
 
 
 const rootNode = document.getElementById('root') as HTMLElement;
 const root = ReactDOM.createRoot(rootNode);
 
-// root.render(<MapGrid/>);
+// root.render(<MapGridRECOVERY2/>);
 root.render(<App />);
-// root.render(<MapGrid />);
+// root.render(<MapGridRECOVERY2 />);
